connection method removed into otdel'niy file random data for adding users into db fixed db name error
heroku mysql addon testing v2 (second try)
