+++ services/mysql/add.user.service.js
@@ -1,21 +1,17 @@
-const mysql = require("mysql2");
+const connection = require("./connection.service");
 
 const addUser = () => {
-    const connection = mysql.createConnection((process.env.PORT) ? {
-        host: "eu-cdbr-west-01.cleardb.com",
-        user: "bedce05604ab0f",
-        database: "terrain_heroes",
-        password: "09dc02a8"
-    } : {
-        host: "localhost",
-        user: "root",
-        database: "terrain_heroes",
-        password: ""
-    });
+    let query = "INSERT INTO `users` (`id`, `login`, `password`, `name`, `age`, `session`) VALUES ( ?, ?, ?, ?, ?);";
 
-    let query = "INSERT INTO `users` (`id`, `login`, `password`, `name`, `age`, `session`) VALUES ( 'Tom', '1234', 'Tom', '24', 'jrRsfKWGF_'), ( 'zul_jin_', 'balthazar', 'Baal', '24', 'kBBsfBXbY2');";
+    let login = randomString(5),
+        password = randomString(5),
+        name = randomString(5),
+        age = Math.floor(Math.random() * 90) + 10,
+        session = randomString(10);
 
-    connection.query(query, function (err, results, fields) {
+    let value = [login, password, name, age, session];
+
+    connection.query(query, value, function (err, results, fields) {
         console.log(results);
     });
 
@@ -24,6 +20,16 @@ const addUser = () => {
             return console.log("Ошибка: " + err.message);
         console.log("Подключение закрыто");
     });
+}
+
+const randomString = (length) => {
+    let result = '',
+        chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_';
+
+    for (let i = 0; i < length; i++)
+        result += chars.charAt(Math.floor(Math.random() * chars.length));
+
+    return result;
 }
 
 module.exports = addUser;
